+++ src/TaskRunner/Commands/ComponentCheckCommands.php
@@ -431,14 +431,11 @@ class ComponentCheckCommands extends AbstractCommands
             // Check if the module is allowed for this profile.
             if (!$allowedInProject && !empty($allowedProfiles)) {
                 $allowedProfiles = array_map('trim', explode(',', $allowedProfiles));
-                // Load the project from the website.
-                $project = Website::projectInformation($projectId);
-                if (array_key_exists('profile', $project)) {
-                    if (in_array($project['profile'], $allowedProfiles)) {
-                        $allowedInProject = true;
-                        $message = "The package $packageName is authorised for the profile {$project['profile']}";
-                        $messageType = 'Packages authorised:';
-                    }
+                $profile = $this->getProjectProfile($projectId);
+                if (in_array($profile, $allowedProfiles)) {
+                    $allowedInProject = true;
+                    $message = "The package $packageName is authorised for the profile $profile";
+                    $messageType = 'Packages authorised:';
                 }
             }
 
@@ -556,26 +553,6 @@ class ComponentCheckCommands extends AbstractCommands
     {
         $packages = [];
         $drupalReleaseHistory = new DrupalReleaseHistory();
-
-        if ($this->isWebsiteInstalled()) {
-            $exec = $this->taskExec($this->getBin('drush') . ' pm:security --format=json')
-                ->setVerbosityThreshold(VerbosityThresholdInterface::VERBOSITY_DEBUG)
-                ->run();
-            if (!empty($exec->getExitCode()) && $exec->getExitCode() !== 3) {
-                $this->io->error(['Failed to run: pm:security', $exec->getMessage()]);
-            } else {
-                $result = trim($exec->getMessage());
-                if (!empty($result) && $result !== '[]') {
-                    $data = json_decode($result, true);
-                    if (!empty($data) && is_array($data)) {
-                        $packages = $data;
-                    }
-                }
-            }
-        } else {
-            $this->writeln('Website not installed, skipping pm:security.');
-        }
-
         $exec = $this->taskExec('composer audit --no-dev --locked --no-scripts --format=json')
             ->setVerbosityThreshold(VerbosityThresholdInterface::VERBOSITY_DEBUG)
             ->run();
@@ -876,6 +853,27 @@ class ComponentCheckCommands extends AbstractCommands
     private function getFailedOrPassed(bool $value): string
     {
         return $value ? 'failed' : 'passed';
+    }
+
+    /**
+     * Load given project from website and return the profile in the production env.
+     *
+     * @param string $projectId
+     *   The project to use in the endpoint.
+     */
+    private function getProjectProfile(string $projectId): string
+    {
+        // Load the project from the website.
+        $project = Website::projectInformation($projectId);
+        // Get the profile from the production environment.
+        if (!empty($project['environments'])) {
+            foreach ($project['environments'] as $env) {
+                if (!empty($env['profile']) && $env['type'] === 'Production') {
+                    return $env['profile'];
+                }
+            }
+        }
+        return '';
     }
 
 }