+++ tests/Features/Commands/DockerCommandsTest.php
@@ -63,7 +63,7 @@ class DockerCommandsTest extends AbstractTest
 
         // Run command.
         $result = $this->runCommand($command);
-
+        $this->debugExpectations($result['output'], $expectations);
         // Assert expectations.
         foreach ($expectations as $expectation) {
             $this->assertDynamic($result['output'], $expectation);
@@ -95,10 +95,10 @@ class DockerCommandsTest extends AbstractTest
 
         // Run command.
         $this->runCommand($command, false);
-
         // Assert expectations.
         foreach ($expectations as $expectation) {
             $content = file_get_contents($this->getSandboxFilepath($expectation['file']));
+            $this->debugExpectations($content, [$expectation]);
             $this->assertDynamic($content, $expectation);
         }
     }